Set material message is now an object.
